# lib/api-zod/src/generated/types/checkin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORBIT-MESH API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Checkin {
  id: number;
  userId: number;
  status: string;
  latitude?: number | null;
  longitude?: number | null;
  createdAt: string;
}